Move cell exec result handler into a seprate class
