+++ src/platform/common/refBool.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 export class RefBool {
     constructor(private val: boolean) {}
